voice/requirements.txt
Flask==2.1.2
requests==2.27.1
openai==1.0.0
Werkzeug==2.0.3
python-dotenv==1.0.1
pydub==0.25.1
ffmpeg-python==0.2.0